re-grouping classes to fit the package conventions.
