+++ tests/models/prf/test_delayed_gain_norm/test_predict_regression_None_.csv
@@ -1,26 +1,51 @@
 ,response_0,response_1,response_2
-0,0.9924389123916626,0.99590015411376953,0.99041515588760376
-1,0.99775135517120361,0.99506223201751709,0.99403363466262817
-2,0.99572187662124634,0.9836198091506958,0.99229151010513306
-3,0.95452463626861572,0.86572670936584473,0.98181372880935669
-4,0.16388653218746185,0.18834501504898071,0.92410159111022949
-5,0.00014042433758731931,0.0026430168654769659,0.59274673461914062
-6,6.7247172275131106e-09,9.234469871444162e-06,0.083379201591014862
-7,2.0076937345698403e-14,9.5619050455297838e-09,0.0029262579046189785
-8,3.6004304397270635e-21,2.8903524539047876e-12,4.8264661018038169e-05
-9,1.3474414326809088e-30,4.6930919233808578e-17,1.6054715956670407e-07
-10,0,8.5088612256863427e-22,3.1244484777204207e-10
-11,0,0,0
-12,0,0,0
-13,0,0,0
-14,9.8722677837153891e-28,1.262707272218415e-15,9.3132772605031278e-09
-15,5.8347129673337014e-20,1.2149711792197593e-11,2.1138957890798338e-06
-16,2.0324247927000494e-13,3.2732504706700638e-08,0.00020472690812312067
-17,2.5143438620034431e-07,6.9622328737750649e-05,0.01745045930147171
-18,0.0021051676012575626,0.013227523304522038,0.2831229567527771
-19,0.55084687471389771,0.44343915581703186,0.82028412818908691
-20,0.98344355821609497,0.93866181373596191,0.96527862548828125
-21,0.99713039398193359,0.99020522832870483,0.98906207084655762
-22,0.99743276834487915,0.99592345952987671,0.99372571706771851
-23,0.98902618885040283,0.99556010961532593,0.9933820366859436
-24,0.73400330543518066,0.98745423555374146,0.98721563816070557
+0,0,0,0
+1,0,0,0
+2,0,0,0
+3,0,0,0
+4,0,0,0
+5,0.17463637888431549,0.97498941421508789,0.92228740453720093
+6,0.79837620258331299,0.99725240468978882,0.99086689949035645
+7,0.97223144769668579,0.99930524826049805,0.99766898155212402
+8,0.99345511198043823,0.9996572732925415,0.99886536598205566
+9,0.99845319986343384,0.9998047947883606,0.99937742948532104
+10,0.99937725067138672,0.99981790781021118,0.99946290254592896
+11,0.99977046251296997,0.99986118078231812,0.99963730573654175
+12,0.99985796213150024,0.99984109401702881,0.99964773654937744
+13,0.99991858005523682,0.99984890222549438,0.99972808361053467
+14,0.9999234676361084,0.99978923797607422,0.99970495700836182
+15,0.99993264675140381,0.99975293874740601,0.99974232912063599
+16,0.99990260601043701,0.9995732307434082,0.99968290328979492
+17,0.99986988306045532,0.99939072132110596,0.99969065189361572
+18,0.99970978498458862,0.99869918823242188,0.99956852197647095
+19,0.99940323829650879,0.99770945310592651,0.99952220916748047
+20,0.99795955419540405,0.9940345287322998,0.99925267696380615
+21,0.99339866638183594,0.98639398813247681,0.99893867969512939
+22,0.96082717180252075,0.94510424137115479,0.99699366092681885
+23,0.77788400650024414,0.78657931089401245,0.99040955305099487
+24,0.094439141452312469,0.13884538412094116,0.86307573318481445
+25,0,0,0
+26,0,0,0
+27,0,0,0
+28,0.0051295487210154533,0.0094749182462692261,0.087068356573581696
+29,0.41361340880393982,0.47327998280525208,0.8767969012260437
+30,0.86907112598419189,0.84415113925933838,0.96993416547775269
+31,0.98008686304092407,0.96293562650680542,0.99135160446166992
+32,0.99484825134277344,0.98566967248916626,0.99546313285827637
+33,0.99871826171875,0.99489659070968628,0.99767196178436279
+34,0.99946427345275879,0.99729412794113159,0.99828708171844482
+35,0.999794602394104,0.99882215261459351,0.99900507926940918
+36,0.99986851215362549,0.99923759698867798,0.99917858839035034
+37,0.99992167949676514,0.99958819150924683,0.99945861101150513
+38,0.99992388486862183,0.99967497587203979,0.99950087070465088
+39,0.99993085861206055,0.99978452920913696,0.99962955713272095
+40,0.99989581108093262,0.99978846311569214,0.99960988759994507
+41,0.99985575675964355,0.99982839822769165,0.9996752142906189
+42,0.9996679425239563,0.99979323148727417,0.9996151328086853
+43,0.99929440021514893,0.9997936487197876,0.99963831901550293
+44,0.99749135971069336,0.99969488382339478,0.99951767921447754
+45,0.99181181192398071,0.99962717294692993,0.99949103593826294
+46,0.9559510350227356,0.99931764602661133,0.99922841787338257
+47,0.80675303936004639,0.9989696741104126,0.99907857179641724
+48,0.32615146040916443,0.99771440029144287,0.99844759702682495
+49,0.055447421967983246,0.99575155973434448,0.99790585041046143
