+++ tests/models/prf/test_delayed_gain_norm/test_predict_regression_impulse_model1_.csv
@@ -1,26 +1,51 @@
 ,response_0,response_1,response_2
-0,0.9924389123916626,0.99590015411376953,0.99041515588760376
-1,0.99775415658950806,0.99506098031997681,0.99403685331344604
-2,0.99572479724884033,0.98357760906219482,0.99229353666305542
-3,0.95436739921569824,0.86413455009460449,0.9817967414855957
-4,0.15512175858020782,0.17552186548709869,0.92375046014785767
-5,1.1232020369789097e-05,0.00092357589164748788,0.58793348073959351
-6,0.00023275369312614202,0.00032644631573930383,0.076727353036403656
-7,0.00023631560907233506,0.00042855768697336316,0.0016304014716297388
-8,0.00023676374985370785,0.00040217165951617062,4.827195152756758e-05
-9,0.00023726973449811339,0.00035110904718749225,0.00018370924226474017
-10,0.00023822084767743945,0.0002830610319506377,0.00019465963123366237
-11,0.00023983842402230948,0.00020869112631771713,0.00019282118591945618
-12,0.00024200761981774122,0.00014018767978996038,0.00018551465473137796
-13,0.00024404548457823694,8.5879095422569662e-05,0.00017154628585558385
-14,0.00024460189160890877,4.8158133722608909e-05,0.00014841457596048713
-15,0.00024183385539799929,2.4817803932819515e-05,9.4543880550190806e-05
-16,0.00023386160319205374,1.0657301572791766e-05,1.9959703422500752e-05
-17,0.00020484988635871559,3.6681605706689879e-05,0.015389883890748024
-18,0.0010164595441892743,0.012928239069879055,0.27915278077125549
-19,0.54557508230209351,0.44356265664100647,0.81989872455596924
-20,0.98343437910079956,0.93879443407058716,0.96531158685684204
-21,0.99713438749313354,0.99023020267486572,0.98908311128616333
-22,0.9974372386932373,0.99593418836593628,0.99373990297317505
-23,0.98903912305831909,0.99557203054428101,0.99339795112609863
-24,0.73319756984710693,0.98748880624771118,0.9872472882270813
+0,0,0,0
+1,0,0,0
+2,0,0,0
+3,0,0,0
+4,0,0,0
+5,0.17506113648414612,0.97695505619049072,0.92417430877685547
+6,0.80005347728729248,0.99922913312911987,0.99283474683761597
+7,0.97417217493057251,1.0012731552124023,0.99963474273681641
+8,0.99541652202606201,1.001612663269043,1.0008200407028198
+9,1.0004211664199829,1.0017567873001099,1.0013298988342285
+10,1.0013366937637329,1.0017561912536621,1.0014029741287231
+11,1.0017322301864624,1.0018055438995361,1.0015847682952881
+12,1.0018080472946167,1.001772403717041,1.0015844106674194
+13,1.0018705129623413,1.001786470413208,1.0016719102859497
+14,1.0018614530563354,1.0017130374908447,1.0016380548477173
+15,1.0018727779388428,1.001683235168457,1.0016824007034302
+16,1.0018253326416016,1.0014876127243042,1.0016113519668579
+17,1.0017955303192139,1.0013117790222168,1.0016270875930786
+18,1.0016131401062012,1.0005977153778076,1.0014920234680176
+19,1.0013101100921631,0.99960589408874512,1.0014538764953613
+20,0.99983739852905273,0.9958498477935791,1.0011705160140991
+21,0.99527156352996826,0.98800623416900635,1.0008586645126343
+22,0.96252948045730591,0.94463026523590088,0.99886912107467651
+23,0.77739977836608887,0.77201765775680542,0.99219930171966553
+24,0.08576541393995285,0.091853581368923187,0.85974001884460449
+25,0.00060935603687539697,0.0059867440722882748,0.0032880676444619894
+26,0.0011311272392049432,0.004767667967826128,0.0038706250488758087
+27,0.0019021106418222189,0.0035523285623639822,0.0042895926162600517
+28,0.00031498135649599135,0.0023040310479700565,0.05527539923787117
+29,0.37645915150642395,0.45272961258888245,0.87297266721725464
+30,0.86412543058395386,0.84226882457733154,0.97117877006530762
+31,0.98157942295074463,0.96451956033706665,0.9932066798210144
+32,0.99674636125564575,0.98755264282226562,0.99737656116485596
+33,1.0006778240203857,0.99684387445449829,0.99961227178573608
+34,1.0014207363128662,0.99924254417419434,1.000224232673645
+35,1.0017549991607666,1.0007787942886353,1.0009533166885376
+36,1.0018174648284912,1.001185417175293,1.0011183023452759
+37,1.0018727779388428,1.0015413761138916,1.0014059543609619
+38,1.0018607378005981,1.0016181468963623,1.0014384984970093
+39,1.0018699169158936,1.0017324686050415,1.0015741586685181
+40,1.0018166303634644,1.0017250776290894,1.0015437602996826
+41,1.0017800331115723,1.0017709732055664,1.0016165971755981
+42,1.0015692710876465,1.0017225742340088,1.0015448331832886
+43,1.0011987686157227,1.0017293691635132,1.0015759468078613
+44,0.99936169385910034,1.0016161203384399,1.0014426708221436
+45,0.99366104602813721,1.0015554428100586,1.0014244318008423
+46,0.95745742321014404,1.0012279748916626,1.0011473894119263
+47,0.80560767650604248,1.0008857250213623,1.0010058879852295
+48,0.31296440958976746,0.99960124492645264,1.0003582239151001
+49,0.043962828814983368,0.99761819839477539,0.9998212456703186
